Refactor and hook EventTarget
