+++ extension/src/webcat/hooks/entry-page.ts
@@ -2,7 +2,7 @@
 // dynamically updated with data and injected into scripts
 // by response.ts
 
-import { eventHook } from "./events";
+import { eventHook, eventTargetHook } from "./events";
 import { wasmHook } from "./wasm";
 import { serviceWorkerHook, sharedWorkerHook, workerHook } from "./workers";
 
@@ -14,4 +14,5 @@ wasmHook(scope, data);
 sharedWorkerHook(scope, data);
 serviceWorkerHook(scope, data);
 workerHook(scope, data);
+eventTargetHook(scope, data);
 eventHook(scope, data);
