+++ test/tests.py
@@ -48,6 +48,8 @@ BAD_WASM = Hook(
     LOGENTRY_ERROR_SHAREDWORKER_CALLBACK,
     LOGENTRY_ERROR_SHAREDWORKER_HANDLER,
     LOGENTRY_ERROR_WORKER_ONERROR,
+    LOGENTRY_ERROR_WORKER_CALLBACK,
+    LOGENTRY_ERROR_WORKER_HANDLER,
 ] = EXPECTED_LOGS = [
     ["inline:",True],
     ["alert.js:",True],
@@ -64,6 +66,8 @@ BAD_WASM = Hook(
     ["Error: error in sharedworker", True, "callback"],
     ["Error: error in sharedworker", True, "handler"],
     ["Error: error in worker", True, "onerror"],
+    ["Error: error in worker", True, "callback"],
+    ["Error: error in worker", True, "handler"],
 ]
 
 EXPECTED_CSP = {
@@ -136,6 +140,8 @@ def setdiff(a: list, b: list):
                 LOGENTRY_ERROR_SHAREDWORKER_CALLBACK,
                 LOGENTRY_ERROR_SHAREDWORKER_HANDLER,
                 LOGENTRY_ERROR_WORKER_ONERROR,
+                LOGENTRY_ERROR_WORKER_CALLBACK,
+                LOGENTRY_ERROR_WORKER_HANDLER,
             ]), [], [],
         NON_FRAME_PATHS-WASM_PATHS-WORKER_PATHS-{"/js/wasm_frame.js"},
         True,
